Add linkage for public names.h helpers
