+++ src/google/protobuf/compiler/command_line_interface_unittest.cc
@@ -65,6 +65,7 @@
 #include "google/protobuf/compiler/command_line_interface.h"
 #include "google/protobuf/compiler/mock_code_generator.h"
 #include "google/protobuf/compiler/subprocess.h"
+#include "google/protobuf/compiler/cpp/names.h"
 #include "google/protobuf/descriptor.h"
 #include "google/protobuf/io/io_win32.h"
 #include "google/protobuf/io/printer.h"
