+++ src/google/protobuf/compiler/java/helpers.cc
@@ -51,7 +51,6 @@
 #include "absl/strings/string_view.h"
 #include "absl/strings/substitute.h"
 #include "google/protobuf/compiler/java/name_resolver.h"
-#include "google/protobuf/compiler/java/names.h"
 #include "google/protobuf/descriptor.pb.h"
 
 // Must be last.
@@ -69,75 +68,6 @@ const char kThickSeparator[] =
     "// ===================================================================\n";
 const char kThinSeparator[] =
     "// -------------------------------------------------------------------\n";
-
-namespace {
-const char* DefaultPackage(Options options) {
-  return options.opensource_runtime ? "" : "com.google.protos";
-}
-
-bool IsReservedName(absl::string_view name) {
-  static const auto& kReservedNames =
-      *new absl::flat_hash_set<absl::string_view>({
-          "abstract",   "assert",       "boolean",   "break",      "byte",
-          "case",       "catch",        "char",      "class",      "const",
-          "continue",   "default",      "do",        "double",     "else",
-          "enum",       "extends",      "final",     "finally",    "float",
-          "for",        "goto",         "if",        "implements", "import",
-          "instanceof", "int",          "interface", "long",       "native",
-          "new",        "package",      "private",   "protected",  "public",
-          "return",     "short",        "static",    "strictfp",   "super",
-          "switch",     "synchronized", "this",      "throw",      "throws",
-          "transient",  "try",          "void",      "volatile",   "while",
-      });
-  return kReservedNames.contains(name);
-}
-
-bool IsForbidden(const std::string& field_name) {
-  // Names that should be avoided (in UpperCamelCase format).
-  // Using them will cause the compiler to generate accessors whose names
-  // collide with methods defined in base classes.
-  // Keep this list in sync with specialFieldNames in
-  // java/core/src/main/java/com/google/protobuf/DescriptorMessageInfoFactory.java
-  static const auto& kForbiddenNames =
-      *new absl::flat_hash_set<absl::string_view>({
-        // java.lang.Object:
-          "Class",
-          // com.google.protobuf.MessageLiteOrBuilder:
-          "DefaultInstanceForType",
-          // com.google.protobuf.MessageLite:
-          "ParserForType",
-          "SerializedSize",
-          // com.google.protobuf.MessageOrBuilder:
-          "AllFields",
-          "DescriptorForType",
-          "InitializationErrorString",
-          "UnknownFields",
-          // obsolete. kept for backwards compatibility of generated code
-          "CachedSize",
-      });
-  return kForbiddenNames.contains(UnderscoresToCamelCase(field_name, true));
-}
-
-std::string FieldName(const FieldDescriptor* field) {
-  std::string field_name;
-  // Groups are hacky:  The name of the field is just the lower-cased name
-  // of the group type.  In Java, though, we would like to retain the original
-  // capitalization of the type name.
-  if (GetType(field) == FieldDescriptor::TYPE_GROUP) {
-    field_name = field->message_type()->name();
-  } else {
-    field_name = field->name();
-  }
-  if (IsForbidden(field_name)) {
-    // Append a trailing "#" to indicate that the name should be decorated to
-    // avoid collision with other names.
-    field_name += "#";
-  }
-  return field_name;
-}
-
-
-}  // namespace
 
 void PrintGeneratedAnnotation(io::Printer* printer, char delimiter,
                               const std::string& annotation_file,
@@ -236,30 +166,6 @@ std::string ToCamelCase(const std::string& input, bool lower_first) {
   return result;
 }
 
-std::string UnderscoresToCamelCase(const FieldDescriptor* field) {
-  return UnderscoresToCamelCase(FieldName(field), false);
-}
-
-std::string UnderscoresToCapitalizedCamelCase(const FieldDescriptor* field) {
-  return UnderscoresToCamelCase(FieldName(field), true);
-}
-
-std::string CapitalizedFieldName(const FieldDescriptor* field) {
-  return UnderscoresToCapitalizedCamelCase(field);
-}
-
-std::string UnderscoresToCamelCase(const MethodDescriptor* method) {
-  return UnderscoresToCamelCase(method->name(), false);
-}
-
-std::string UnderscoresToCamelCaseCheckReserved(const FieldDescriptor* field) {
-  std::string name = UnderscoresToCamelCase(field);
-  if (IsReservedName(name)) {
-    return name + "_";
-  }
-  return name;
-}
-
 // Names that should be avoided as field names in Kotlin.
 // All Kotlin hard keywords are in this list.
 bool IsForbiddenKotlin(absl::string_view field_name) {
@@ -305,53 +211,11 @@ std::string FileClassName(const FileDescriptor* file, bool immutable) {
   return ClassNameResolver().GetFileClassName(file, immutable);
 }
 
-std::string FileJavaPackage(const FileDescriptor* file, bool immutable,
-                            Options options) {
-  std::string result;
-
-  if (file->options().has_java_package()) {
-    result = file->options().java_package();
-  } else {
-    result = DefaultPackage(options);
-    if (!file->package().empty()) {
-      if (!result.empty()) result += '.';
-      result += file->package();
-    }
-  }
-
-  return result;
-}
-
-std::string FileJavaPackage(const FileDescriptor* file, Options options) {
-  return FileJavaPackage(file, true /* immutable */, options);
-}
-
 std::string JavaPackageToDir(std::string package_name) {
   std::string package_dir = absl::StrReplaceAll(package_name, {{".", "/"}});
   if (!package_dir.empty()) package_dir += "/";
   return package_dir;
 }
-
-std::string ClassName(const Descriptor* descriptor) {
-  ClassNameResolver name_resolver;
-  return name_resolver.GetClassName(descriptor, true);
-}
-
-std::string ClassName(const EnumDescriptor* descriptor) {
-  ClassNameResolver name_resolver;
-  return name_resolver.GetClassName(descriptor, true);
-}
-
-std::string ClassName(const ServiceDescriptor* descriptor) {
-  ClassNameResolver name_resolver;
-  return name_resolver.GetClassName(descriptor, true);
-}
-
-std::string ClassName(const FileDescriptor* descriptor) {
-  ClassNameResolver name_resolver;
-  return name_resolver.GetClassName(descriptor, true);
-}
-
 
 std::string ExtraMessageInterfaces(const Descriptor* descriptor) {
   std::string interfaces = "// @@protoc_insertion_point(message_implements:" +
