+++ src/google/protobuf/compiler/java/names.h
@@ -96,6 +96,31 @@ std::string FileJavaPackage(const FileDescriptor* descriptor,
 //   Capitalized camel case name field name.
 std::string CapitalizedFieldName(const FieldDescriptor* descriptor);
 
+// Returns:
+//   Converts a name to camel-case. If cap_first_letter is true, capitalize the
+//   first letter.
+std::string UnderscoresToCamelCase(const std::string& name,
+                                   bool cap_first_letter);
+// Requires:
+//   field != NULL
+// Returns:
+//   Converts the field's name to camel-case, e.g. "foo_bar_baz" becomes
+//   "fooBarBaz" or "FooBarBaz", respectively.
+std::string UnderscoresToCamelCase(const FieldDescriptor* field);
+
+// Requires:
+//   method != NULL
+// Returns:
+//   Similar, but for method names.  (Typically, this merely has the effect
+//   of lower-casing the first letter of the name.)
+std::string UnderscoresToCamelCase(const MethodDescriptor* method);
+
+// Requires:
+//   field != NULL
+// Returns:
+//   Same as UnderscoresToCamelCase, but checks for reserved keywords
+std::string UnderscoresToCamelCaseCheckReserved(const FieldDescriptor* field);
+
 }  // namespace java
 }  // namespace compiler
 }  // namespace protobuf
