+++ src/google/protobuf/compiler/objectivec/BUILD.bazel
@@ -9,6 +9,10 @@ load("//build_defs:cpp_opts.bzl", "COPTS")
 cc_library(
     name = "names",
     hdrs = ["names.h"],
+    srcs = [
+        "names.cc",
+        "objectivec_nsobject_methods.h",
+    ],
     copts = COPTS,
     include_prefix = "google/protobuf/compiler/objectivec",
     visibility = ["//visibility:public"],
@@ -24,7 +28,6 @@ cc_library(
         "objectivec_field.cc",
         "objectivec_file.cc",
         "objectivec_generator.cc",
-        "objectivec_helpers.cc",
         "objectivec_map_field.cc",
         "objectivec_message.cc",
         "objectivec_message_field.cc",
@@ -42,7 +45,6 @@ cc_library(
         "objectivec_map_field.h",
         "objectivec_message.h",
         "objectivec_message_field.h",
-        "objectivec_nsobject_methods.h",
         "objectivec_oneof.h",
         "objectivec_options.h",
         "objectivec_primitive_field.h",
