+++ src/google/protobuf/compiler/php/BUILD.bazel
@@ -9,10 +9,15 @@ load("//build_defs:cpp_opts.bzl", "COPTS")
 cc_library(
     name = "names",
     hdrs = ["names.h"],
+    srcs = ["names.cc"],
     copts = COPTS,
     include_prefix = "google/protobuf/compiler/php",
     visibility = ["//visibility:public"],
-    deps = ["//src/google/protobuf:protobuf_nowkt"],
+    deps = [
+        "//src/google/protobuf:protobuf_nowkt",
+        "//src/google/protobuf/compiler:code_generator",
+        "@com_google_absl//absl/strings",
+    ],
 )
 
 cc_library(
